feat(LLD): more MAD alignment
