+++ apps/ledger-live-desktop/src/newArch/features/ModularDrawer/components/SelectAccount.tsx
@@ -1,108 +1,48 @@
-import React, { useMemo } from "react";
-import { CryptoOrTokenCurrency } from "@ledgerhq/types-cryptoassets";
-import { AccountList, CardButton } from "@ledgerhq/react-ui/pre-ldls";
+import React from "react";
 import { track } from "~/renderer/analytics/segment";
 import TrackPage from "~/renderer/analytics/TrackPage";
+import { Box, Flex } from "@ledgerhq/react-ui/index";
 import { Account, AccountLike } from "@ledgerhq/types-live";
-import { Icons } from "@ledgerhq/react-ui/index";
-import { useTranslation } from "react-i18next";
-import { getAccountTuplesForCurrency } from "~/renderer/components/PerCurrencySelectAccount/state";
-import { useGetAccountIds } from "@ledgerhq/live-common/wallet-api/react";
-import { useSelector } from "react-redux";
-import { accountsSelector } from "~/renderer/reducers/accounts";
-import { Observable } from "rxjs";
-import { WalletAPIAccount } from "@ledgerhq/live-common/wallet-api/types";
-import {
-  getBalanceHistoryWithCountervalue,
-  getPortfolioCount,
-} from "@ledgerhq/live-countervalues/lib-es/portfolio";
-import { useCountervaluesState } from "@ledgerhq/live-countervalues-react";
-import { counterValueCurrencySelector } from "~/renderer/reducers/settings";
-import { formatCurrencyUnit } from "@ledgerhq/coin-framework/lib-es/currencies/formatCurrencyUnit";
-import BigNumber from "bignumber.js";
-import { getTagDerivationMode } from "@ledgerhq/coin-framework/lib-es/derivation";
+import { AccountList, Account as DetailedAccount } from "@ledgerhq/react-ui/pre-ldls/index";
 
-type Props = {
-  asset: CryptoOrTokenCurrency;
+type SelectAccountProps = {
   onAccountSelected: (account: AccountLike, parentAccount?: Account) => void;
   source: string;
   flow: string;
-  accounts$?: Observable<WalletAPIAccount[]>;
+  detailedAccounts: DetailedAccount[];
 };
 
-const useDetailedAccounts = (
-  asset: CryptoOrTokenCurrency,
-  accounts$?: Observable<WalletAPIAccount[]>,
-) => {
-  const accountIds = useGetAccountIds(accounts$);
-  const nestedAccounts = useSelector(accountsSelector);
-  const to = useSelector(counterValueCurrencySelector);
-  const state = useCountervaluesState();
+const CURRENT_PAGE = "Modular Account Selection";
+const ROW_HEIGHT = 64;
+const SPACING = 8;
+const TWO_ROWS_HEIGHT = 2 * ROW_HEIGHT;
+const INPUT_HEIGHT = 40;
+const EXTRA_BOTTOM_MARGIN = TWO_ROWS_HEIGHT + INPUT_HEIGHT + SPACING;
 
-  const accounts = getAccountTuplesForCurrency(asset, nestedAccounts, false, accountIds);
-
-  return useMemo(
-    () =>
-      accounts.map(({ account }) => {
-        const count = getPortfolioCount([account], "day");
-
-        const balance = formatCurrencyUnit(account.currency.units[0], account.balance, {
-          showCode: true,
-        });
-
-        const { history } = getBalanceHistoryWithCountervalue(account, "day", count, state, to);
-        const { countervalue } = history[history.length - 1];
-        const fiatValue = formatCurrencyUnit(to.units[0], BigNumber(countervalue || 0), {
-          showCode: true,
-        });
-
-        const address = `${account.seedIdentifier.slice(0, 5)}...${account.seedIdentifier.slice(-5)}`;
-        const protocol = getTagDerivationMode(account.currency, account.derivationMode); // TODO consider capital letters
-
-        return {
-          name: account.currency.name,
-          id: account.currency.id,
-          ticker: account.currency.ticker,
-          balance,
-          fiatValue,
-          protocol: protocol || "",
-          address,
-        };
-      }),
-    [accounts, state, to],
-  );
-};
-
-export const SelectAccount = ({ asset, accounts$, onAccountSelected, source, flow }: Props) => {
-  const detailedAccounts = useDetailedAccounts(asset, accounts$);
-  const { t } = useTranslation();
-
-  const onAddAccountClick = () => {
-    // TODO: To be implemented in LIVE-17272
-    track("button_clicked", {
-      button: "Add a new account",
-      page: "Modular Account Selection",
-      flow,
-    });
-  };
-
+export const SelectAccount = ({
+  detailedAccounts,
+  onAccountSelected,
+  source,
+  flow,
+}: SelectAccountProps) => {
   const onAccountClick = (networkId: string) => {
     track("account_clicked", { currency: networkId, page: "Modular Account Selection", flow });
+    // TODO to be implemented as part of LIVE-18127
     onAccountSelected({} as AccountLike, {} as Account);
   };
 
   return (
-    <>
-      <TrackPage category={source} name="Modular Account Selection" flow={flow} />
-      <CardButton
-        onClick={onAddAccountClick}
-        title={t("drawers.selectAccount.addAccount")}
-        iconRight={<Icons.Plus size="S" />}
-        variant="dashed"
-      />
-      <div style={{ flex: "1 1 auto", width: "100%", padding: "16px 0 0 0" }}>
+    <Box style={{ height: "100%", display: "flex", flexDirection: "column" }}>
+      <TrackPage category={source} name={CURRENT_PAGE} flow={flow} />
+      <Flex
+        style={{
+          flex: "1",
+          overflow: "auto",
+          paddingBottom: `${EXTRA_BOTTOM_MARGIN}px`,
+        }}
+      >
         <AccountList accounts={detailedAccounts} onClick={onAccountClick} />
-      </div>
-    </>
+      </Flex>
+    </Box>
   );
 };
