+++ apps/ledger-live-desktop/src/newArch/features/ModularDrawer/components/SelectAccountFlow/components/index.tsx
@@ -1,2 +1,2 @@
-export * from "./Header";
 export * from "./StyledComponents";
+export * from "./AccountSelectionStep";
