+++ apps/ledger-live-desktop/src/newArch/features/ModularDrawer/components/SelectAssetFlow/components/index.tsx
@@ -1,4 +1,3 @@
 export * from "./AssetSelectionStep";
 export * from "./NetworkSelectionStep";
-export * from "./Header";
 export * from "./StyledComponents";
