+++ libs/ui/packages/react/src/pre-ldls/components/AccountItem/AccountItem.tsx
@@ -26,6 +26,7 @@ const Wrapper = styled.div`
     "spacing-xxxs",
     "spacing-xxs",
     "spacing-xs",
+    "radius-s",
     "colors-content-default-default",
     "colors-surface-transparent-hover",
     "colors-surface-transparent-pressed",
@@ -35,6 +36,7 @@ const Wrapper = styled.div`
   display: flex;
   padding: var(--spacing-xs) var(--spacing-xxs);
   cursor: pointer;
+  border-radius: var(--radius-s, 8px);
   justify-content: space-between;
   align-items: center;
